Refactor timeline rail types into Rail directory
Move and reorganize timeline rail types into a new Rail directory. Several DeviceRail-related files were renamed/moved (Dedup, DeviceRail+Init, DeviceRail+Merge, DeviceRail.swift) and tests stubs updated accordingly. The large DeviceRail+Sample.swift was removed and a compact sample extension was added to DeviceRail.swift. New lightweight UI model types were introduced: InstallRail, LaunchRail and SessionRail (each with sample helpers). Redundant rail structs were removed from Records/Install, Records/Launch and Records/Session to keep record models separate from UI rail models. This cleans up the project structure and separates UI "rail" view models from the underlying record definitions.
