+++ Sources/Scout/UI/Timeline/Rail/DeviceRail+Init.swift
