+++ Sources/Scout/UI/Timeline/Rail/DeviceRail+Merge.swift
