+++ Tests/ScoutTests/Stubs/Rail/DeviceRail+Stub.swift
